% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distUtils.R
\name{dist.create}
\alias{dist.create}
\title{Create an instance of \code{\link[stats]{dist}}}
\usage{
dist.create(distances, names)
}
\arguments{
\item{distances}{the distance vector}

\item{names}{the vector of names}
}
\value{
an instance \code{\link[stats]{dist}}
}
\description{
Provide a distance matrix and names and then create a get an
  instance of \code{\link[stats]{dist}} in return.
}
