% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distUtils.R
\name{dist.get}
\alias{dist.get}
\title{Compute the Distance between Two Objects based on a
  \code{\link[stats]{dist}} Object}
\usage{
dist.get(distObj, i, j, n = attr(x = distObj, which = "Size", exact = TRUE))
}
\arguments{
\item{distObj}{the instance of \code{\link[stats]{dist}}}

\item{i}{the first row or column index in the distance matrix}

\item{j}{the second row or column index in the distance matrix}

\item{n}{the number of rows/columns of the distance matrix}
}
\value{
the distance between the objects at indices \code{i} and \code{j}
}
\description{
Compute the index into a \code{\link[stats]{dist}} object based
  on two indexes of objects whose distance should be computed and return the
  corresponding distance.
}
