% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distUtils.R
\name{dist.ij}
\alias{dist.ij}
\title{Compute the Indices of the Two Objects Corresponding to a
  \code{\link[stats]{dist}} Object Index}
\usage{
dist.ij(index, n)
}
\arguments{
\item{index}{the index}

\item{n}{the number of rows/columns of the distance matrix}
}
\value{
a vector with the two indices
}
\description{
Compute the row and column indices of the two objects
  corresponding to a \code{\link[stats]{dist}} object index. See
  \url{http://stackoverflow.com/questions/9879608}.
}
