% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distUtils.R
\name{dist.index}
\alias{dist.index}
\title{Compute the Index into a \code{\link[stats]{dist}} Object}
\usage{
dist.index(i, j, n)
}
\arguments{
\item{i}{the first row or column index in the distance matrix}

\item{j}{the second row or column index in the distance matrix}

\item{n}{the number of rows/columns of the distance matrix}
}
\value{
\code{0L} if \code{i == j}, else the index in the
  \code{\link[stats]{dist}} object
}
\description{
Compute the index into a \code{\link[stats]{dist}} object based
  on two indexes of objects whose distance should be computed. See
  \url{http://stackoverflow.com/questions/9879608}.
}
