% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distUtils.R
\name{dist.indexes}
\alias{dist.indexes}
\title{Generate a List of Index Pairs that can be used to Construct a
  \code{\link[stats]{dist}} Object}
\usage{
dist.indexes(n)
}
\arguments{
\item{n}{the number of objects to compare}
}
\value{
a list of object indexes that need to be compared, exactly in the
  same style and order as in \code{\link[stats]{dist}} objects
}
\description{
A list with the index pairs to be used to construct a distance
  matrix of \code{n} rows and columns is generated.
}
