% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedDistance.R
\name{Models.dist.aggregate}
\alias{Models.dist.aggregate}
\title{Create a Distance Matrix Corresponding to the Aggregated Result of the
  Model Distances}
\usage{
Models.dist.aggregate(models, distance = RegressionResult.dist.default,
  aggregate = mean, cores = 1L, logging = FALSE)
}
\arguments{
\item{models}{the models records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{aggregate}{the function used for joining the models created by
\code{distance}}

\item{cores}{the number of CPU cores to use (ignored)}

\item{logging}{the logging setup, see \code{\link[utilizeR]{makeLogger}}}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the aggregated
  distances between the Models.
}
\seealso{
Models.dist.mean

Models.dist.median
}
