% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distDefault.R
\name{Models.dist.default}
\alias{Models.dist.default}
\title{Create a Distance Matrix of Fitted Models}
\usage{
Models.dist.default(models, distance = RegressionResult.dist.default,
  cores = 1L)
}
\arguments{
\item{models}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{cores}{the number of CPU cores to use}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the distances
  between the models. For each model. This method represents the default
  approach for that, which current is \code{\link{Models.dist.rank.mean}}.
}
\seealso{
Models.dist.rank.mean
}
