% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoinedDistance.R
\name{Models.dist.join}
\alias{Models.dist.join}
\title{Create a Distance Matrix Corresponding to the Joined Result of the
  Model Distances}
\usage{
Models.dist.join(models, distance = RegressionResult.dist.default,
  join = mean, cores = 1L)
}
\arguments{
\item{models}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{join}{the function used for joining the models created by
\code{distance}}

\item{cores}{the number of CPU cores to use (ignored)}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the joined
  distances between the Models.
}
\seealso{
Models.dist.mean

Models.dist.median
}
