% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedDistance.R
\name{Models.dist.mean}
\alias{Models.dist.mean}
\title{Create a Distance Matrix Corresponding to the Mean of the Model
  Distances}
\usage{
Models.dist.mean(models, distance = RegressionResult.dist.default,
  cores = 1L)
}
\arguments{
\item{models}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{cores}{the number of CPU cores to use (ignored)}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the mean
  distances between the Models.
}
\seealso{
Models.dist.aggregate

Models.dist.median
}
