% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedRankDistance.R
\name{Models.dist.rank.global.mean}
\alias{Models.dist.rank.global.mean}
\title{Create a Distance Matrix Corresponding to the Mean Ranks over all
  Model Distances}
\usage{
Models.dist.rank.global.mean(models, distance = RegressionResult.dist.default,
  cores = 1L, logging = FALSE)
}
\arguments{
\item{models}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{cores}{the number of CPU cores to use}

\item{logging}{the logging setup, see \code{\link[utilizeR]{makeLogger}}}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
First we compute the distances between all models not in the
  same \code{\link[processMineR.models]{Models}} set. These distances are
  then ranked using the normalizing ranking method
  \code{\link[distanceR]{rank.dist}}. The distance of one
  \code{\link[processMineR.models]{Models}} set to another model set is then
  the mean of the corresponding ranks.
}
\seealso{
Models.dist.rank.aggregate

Models.dist.rank.local.mean
}
