% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoinedRankDistance.R
\name{Models.dist.rank.join}
\alias{Models.dist.rank.join}
\title{Create a Distance Matrix Corresponding to the Joined Result of the
  Ranked Model Distances}
\usage{
Models.dist.rank.join(models, distance = RegressionResult.dist.default,
  join = mean, cores = 1L)
}
\arguments{
\item{models}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{join}{the function used for joining the models created by
\code{distance}}

\item{cores}{the number of CPU cores to use}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the joined
  ranked distances between the models. For each model, we rank all the other
  models according to their distance. This should be robust against some odd
  outliers.
}
\seealso{
Models.dist.rank.mean

Models.dist.rank.median
}
