% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedRankDistance.R
\name{Models.dist.rank.local.mean}
\alias{Models.dist.rank.local.mean}
\title{Create a Distance Matrix Corresponding to the Mean Ranks of the
  Distances from Each Model to any other Model}
\usage{
Models.dist.rank.local.mean(models, distance = RegressionResult.dist.default,
  cores = 1L, logging = FALSE)
}
\arguments{
\item{models}{the models records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{cores}{the number of CPU cores to use}

\item{logging}{the logging setup, see \code{\link[utilizeR]{makeLogger}}}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
For each model \code{M} we compute the distances to all other
  models. These distances are then ranked using the normalizing ranking
  method \code{\link[distanceR]{rank.dist}}. The distance of one models set
  to another model set is then the mean of the corresponding ranks.
}
\seealso{
Models.dist.rank.aggregate

Models.dist.rank.global.mean
}
