% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoinedRankDistance.R
\name{Models.dist.rank.mean}
\alias{Models.dist.rank.mean}
\title{Create a Distance Matrix Corresponding to the Mean of the Model
  Distances}
\usage{
Models.dist.rank.mean(results, distance = RegressorResult.dist.default,
  cores = 1L)
}
\arguments{
\item{results}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{cores}{the number of CPU cores to use}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the mean
  distances between the models.
}
\seealso{
Models.dist.rank.join

Models.dist.rank.median
}
