% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoinedTotalRankDistance.R
\name{Models.dist.totalRank.join}
\alias{Models.dist.totalRank.join}
\title{Create a Distance Matrix Corresponding to the Joined Result of the
  Totally Ranked Model Distances}
\usage{
Models.dist.totalRank.join(models, distance = RegressionResult.dist.default,
  join = mean, cores = 1L)
}
\arguments{
\item{models}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{join}{the function used for joining the models created by
\code{distance}}

\item{cores}{the number of CPU cores to use}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the joined
  totally ranked distances between the models. We first compute all the
  distances between all the models and then rank them. These ranks are then
  used instead of the actual distances to build the distance matrix.
}
\seealso{
Models.dist.totalRank.mean

Models.dist.totalRank.median
}
