% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoinedTotalRankDistance.R
\name{Models.dist.totalRank.mean}
\alias{Models.dist.totalRank.mean}
\title{Create a Distance Matrix Corresponding to the Mean of the Model
  Distances}
\usage{
Models.dist.totalRank.mean(models, distance = RegressionResult.dist.default,
  cores = 1L)
}
\arguments{
\item{models}{the \code{\link[processMineR.models]{Models}} records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{cores}{the number of CPU cores to use}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the mean
  distances between the models.
}
\seealso{
Models.dist.totalRank.join

Models.dist.totalRank.median
}
