% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceDefault.R
\name{RegressionResult.dist.default}
\alias{RegressionResult.dist.default}
\title{The default \code{\link[regressoR]{RegressionResult}} Distance Metric}
\usage{
RegressionResult.dist.default(a, b)
}
\arguments{
\item{a}{the first instance of \code{\link{RegressionResult}}}

\item{b}{the second instance of \code{\link{RegressionResult}}}
}
\value{
a distance value in \code{[0,+Inf]}
}
\description{
The default distance metric for two instances \code{a} and
  \code{b}  of \code{\link[regressoR]{RegressionResult}}.
}
