% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceQuality.R
\name{RegressionResult.dist.quality}
\alias{RegressionResult.dist.quality}
\title{A \code{\link{RegressionResult}} Distance Metric based on Plain
  Quality}
\usage{
RegressionResult.dist.quality(a, b)
}
\arguments{
\item{a}{the first instance of \code{\link{RegressionResult}}}

\item{b}{the second instance of \code{\link{RegressionResult}}}
}
\value{
a distance value in \code{[0,+Inf]}
}
\description{
The distance of two instances \code{a} and \code{b} of
  \code{\link{RegressionResult}} is based on how well the models represent
  the points of the other's source data.
}
