% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedDistance.R
\name{Models.dist.median}
\alias{Models.dist.median}
\title{Create a Distance Matrix Corresponding to the Median of the Model
  Distances}
\usage{
Models.dist.median(models, distance = RegressionResult.dist.default,
  cores = 1L, logging = FALSE)
}
\arguments{
\item{models}{the models records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{cores}{the number of CPU cores to use (ignored)}

\item{logging}{the logging setup, see \code{\link[utilizeR]{makeLogger}}}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the median
  distances between the Models.
}
\seealso{
Models.dist.aggregate

Models.dist.mean
}
