% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedRankDistance.R
\name{Models.dist.rank.aggregate}
\alias{Models.dist.rank.aggregate}
\title{Create a Distance Matrix Corresponding to the Aggregated Result of the
  Ranked Model Distances}
\usage{
Models.dist.rank.aggregate(models, distance = RegressionResult.dist.default,
  rank.all = rank.dist, rank.fromSingle = identity, aggregate = mean,
  cores = 1L, logging = FALSE)
}
\arguments{
\item{models}{the Models records}

\item{distance}{the
\code{\link[regressoR]{RegressionResult}}-to-\code{\link[regressoR]{RegressionResult}}
 distance metric}

\item{rank.all}{the ranking to be applied to all distances}

\item{rank.fromSingle}{the ranking to be applied to all the distances from
one specific sample to the other samples}

\item{aggregate}{the function used for joining the models created by
\code{distance}}

\item{cores}{the number of CPU cores to use}

\item{logging}{the logging setup, see \code{\link[utilizeR]{makeLogger}}}
}
\value{
an instance of \code{\link[stats]{dist}}
}
\description{
A distance matrix is constructed which represents the aggregated
  ranked distances between the models. This should be robust against some odd
  outliers.
}
\seealso{
Models.dist.rank.mean

Models.dist.rank.median
}
